added more fixtures

+++ hhpawapi/models/item.py
@@ -2,4 +2,4 @@ from django.db import models
 
 class Item(models.Model):
     name = models.CharField(max_length=50)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=5, decimal_places=2)
