+++ hhpawapi/models/revenue.py
@@ -3,8 +3,8 @@ from .order import Order
 
 class Revenue(models.Model):
     order = models.ForeignKey(Order, on_delete=models.CASCADE)
-    order_amount = models.IntegerField()
-    tip_amount = models.IntegerField()
+    order_amount = models.DecimalField(max_digits=5, decimal_places=2)
+    tip_amount = models.DecimalField(max_digits=5, decimal_places=2)
     payment_type = models.CharField(max_length=50)
     date = models.DateTimeField(auto_now=False)
